Implementando post, put e delete no controller

+++ backend/backend/Model/Despesa.cs
@@ -6,7 +6,7 @@ namespace backend.Model
     public class Despesa
     {
         [Key]
-        public long idDespesa { get; set; }
+        public int idDespesa { get; set; }
         public string dataCadastro { get; set; }
         public string descricao { get; set; }
         public int idTipo {get; set;}
